in progress

+++ frontend/works_shg_app/lib/models/employee/mb/mb_inbox_response.dart
@@ -194,7 +194,7 @@ class Measure with _$Measure {
       @JsonKey(name: 'id')
           String? id,
       @JsonKey(name: 'currentValue')
-          int? currentValue,
+          double? currentValue,
       @JsonKey(name: 'cumulativeValue')
           int? cumulativeValue,
       @JsonKey(name: 'height')
